# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jet on Set API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Personal Information"
                ],
                "summary": "Display a listing of the resource.",
                "description": "User data Page",
                "operationId": "522a0d81bcd0708b222a1efda64e34f1",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Personal Information"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Update profile data",
                "operationId": "a796577534095369929ad6f3e6fea1dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "example": "first name"
                                    },
                                    "last_name": {
                                        "description": "Last Name",
                                        "type": "string",
                                        "example": "last name"
                                    },
                                    "phone_number": {
                                        "description": "Phone Number",
                                        "type": "string",
                                        "example": "+44444444444"
                                    },
                                    "date_of_birth": {
                                        "description": "Date of birth",
                                        "type": "string",
                                        "example": "11/12/1986"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string",
                                        "example": "larapoints123"
                                    },
                                    "country": {
                                        "description": "Country",
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "city": {
                                        "description": "City",
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "postcode": {
                                        "description": "Postcode",
                                        "type": "string",
                                        "example": "333333333"
                                    },
                                    "billing_address": {
                                        "description": "Billing Address",
                                        "type": "string",
                                        "example": "Address"
                                    },
                                    "billing_country": {
                                        "description": "Billing Country",
                                        "type": "string",
                                        "example": "Country"
                                    },
                                    "billing_city": {
                                        "description": "Billing City",
                                        "type": "string",
                                        "example": "City"
                                    },
                                    "billing_state": {
                                        "description": "Billing State",
                                        "type": "string",
                                        "example": "State"
                                    },
                                    "billing_postcode": {
                                        "description": "Billing Postcode",
                                        "type": "string",
                                        "example": "Postcode"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}